signout works as screen. nav renders right when user logged in
