+++ client/src/components/index.js
@@ -3,6 +3,6 @@ import Product from './Product/Product';
 import Search from './Search/Search';
 import Sort from './Sort/Sort';
 import FeaturedProduct from './FeaturedProduct/FeaturedProduct';
-import SignOut from './SignOut/SignOut'
+import SignOut from '../screens/SignOut/SignOut';
 
 export { Layout, Product, Search, Sort, FeaturedProduct, SignOut }
